+++ cutile/tests/kernel_compiler.rs
@@ -8,6 +8,7 @@ use cutile::cutile_compiler::cuda_tile_runtime_utils::{
     serialize_tile_ir_bytecode, tileiras_fingerprint, TileirasOptions,
 };
 use cutile::jit_cache::l2_key;
+use cutile::tile_kernel::CompileOptions;
 
 mod common;
 
@@ -91,7 +92,7 @@ fn kernel_compiler_l2_cache_key_matches_runtime_derivation() {
             &bytecode,
             version,
             "sm_120",
-            &TileirasOptions::default(),
+            &TileirasOptions::from_compile_options(&CompileOptions::default()),
             tileiras_fingerprint(),
         );